WHY: HOW TO USE: * New documentation
fnb_learn.8.in:
* Correct spelling, grammar
* Add new information

bayes.h:
bayes.c:
* Change some uint_64t and int64_t to their 32bit counter parts as the 64 bit version isn't needed.
* Change some unsigned to signed and/or vice versa as appropriate.
* Make more functions static and/or inline.

hyperspace.h:
hyperspace.c:
* Change some uint_64t and int64_t to their 32bit counter parts as the 64 bit version isn't needed.
* Change some unsigned to signed and/or vice versa as appropriate.

html.h:
html.c:
* Default to using new PATRICIA TRIE structure in computeOSBHashes. This causes about 4 times the memory to be used for 3-60 ms. However, this speeds up execution at least some what in all but a few cases (such as slashdot.org where it slows things down by 1-2 ms). In many cases, including training, this causes marked speed ups!
* Better order functions for better cache hits.
* Regex clean up leading to reasonable speed up for small documents and huge gains in speed for large documents.
* Various small optimizations.
* Make more functions static and/or inline.
* Fix memory leak that was introduced when we converted to libicu for parsing in computeOSBHashes. This only happened when the maximum number of hashes was exceeded.

fhs_judge.c:
fnb_judge.c:
* Clean up timing code for debugging and optimization.

fnb_learn.c:
* Remove fnb files that should be automatically. No longer require user to do it.
* Use fnb file as preload before removing it (helps in marked speed up in training).

srv_classify_image.c:
* Fix printf formatting.

+++ services/classify/hyperspace.c
@@ -336,9 +336,9 @@ uint16_t hash;
 }
 #endif
 
-static int64_t HSBinarySearch(HashListExt *hashes_list, int64_t start, int64_t end, uint64_t key)
+static int32_t HSBinarySearch(HashListExt *hashes_list, int32_t start, int32_t end, uint64_t key)
 {
-int64_t mid=0;
+int32_t mid=0;
 	if(start > end)
 		return -1;
 	mid = start + ((end - start) / 2);
@@ -395,7 +395,7 @@ int loadHyperSpaceCategory(const char *fhs_name, const char *cat_name)
 {
 int fhs_file;
 uint16_t i, j, z, shortcut=0, offsetPos=2;
-int64_t BSRet=-1;
+int32_t BSRet=-1;
 int64_t offsets[HS_OFFSET_MAX+1];
 HTMLFeature *docHashes;
 FHS_HEADERv1 header;
@@ -766,7 +766,7 @@ HTMLClassification doHSPrepandClassify(HashList *toClassify)
 {
 uint32_t i, j;
 uint32_t **categories = NULL;
-int64_t BSRet = -1;
+int32_t BSRet = -1;
 HTMLClassification data = { .primary_name = NULL, .primary_probability = 0.0, .primary_probScaled = 0.0, .secondary_name = NULL, .secondary_probability = 0.0, .secondary_probScaled = 0.0  };;
 
 	if(HSCategories.used < 2) return data; // We must have at least two categories loaded or it is pointless to run
